extracted to common hook to avoid code duplication
